requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
pandas==1.3.0
django-cors-headers==3.7.0
